[PEx] Refactoring and formatting changes

+++ Src/PRuntimes/PExRuntime/src/main/java/pex/runtime/scheduler/explicit/SchedulerStatistics.java
@@ -19,7 +19,7 @@ public class SchedulerStatistics {
     /**
      * Total steps
      */
-    public int totalSteps = 0;
+    public long totalSteps = 0;
 
     /**
      * Total number of states visited
